requirements.txt
Flask==3.0.3
google-generativeai>=0.8.0
python-dotenv==1.0.1
